// oppabControlliClient/src/app/model/utenteGruppo.ts
import { Gruppo } from './gruppo';
import { Utente } from './utente';

export class UtenteGruppo {
  id: number;
  idRel: number;
  utente: Utente;
  gruppo: Gruppo;
/*
  nomeUtente: string;
  cognomeUtente: string;
  usernameUtente: string;
  emailUtente: string;
  dominioUtente: string;

  nomeGruppo: string;
  descrizioneITGruppo: string;
  descrizioneDEGruppo: string;*/

}